Move Institution to model project
